+++ src/DataContainer/LayoutContainer.php
@@ -14,24 +14,31 @@ use Contao\CoreBundle\ServiceAnnotation\Callback;
 use Contao\DataContainer;
 use Contao\LayoutModel;
 use Contao\Message;
+use HeimrichHannot\EncoreBundle\Collection\EntryCollection;
+use HeimrichHannot\EncoreBundle\Exception\NoEntrypointsException;
 use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class LayoutContainer
 {
-    protected array           $bundleConfig;
-    protected ContaoFramework $contaoFramework;
-    private RequestStack      $requestStack;
-    private ScopeMatcher      $scopeMatcher;
+    protected array             $bundleConfig;
+    protected ContaoFramework   $contaoFramework;
+    private RequestStack        $requestStack;
+    private ScopeMatcher        $scopeMatcher;
+    private EntryCollection     $entryCollection;
+    private TranslatorInterface $translator;
 
     /**
      * LayoutContainer constructor.
      */
-    public function __construct(array $bundleConfig, ContaoFramework $contaoFramework, RequestStack $requestStack, ScopeMatcher $scopeMatcher)
+    public function __construct(array $bundleConfig, ContaoFramework $contaoFramework, RequestStack $requestStack, ScopeMatcher $scopeMatcher, EntryCollection $entryCollection, TranslatorInterface $translator)
     {
         $this->bundleConfig = $bundleConfig;
         $this->contaoFramework = $contaoFramework;
         $this->requestStack = $requestStack;
         $this->scopeMatcher = $scopeMatcher;
+        $this->entryCollection = $entryCollection;
+        $this->translator = $translator;
     }
 
     /**
@@ -41,17 +48,31 @@ class LayoutContainer
     {
         $request = $this->requestStack->getCurrentRequest();
 
-        if (!$request
-            || !$dc
-            || !$this->scopeMatcher->isBackendRequest($request)
-            || !isset($this->bundleConfig['use_contao_template_variables'])
-            || true !== $this->bundleConfig['use_contao_template_variables']
-            || !($layout = $this->contaoFramework->getAdapter(LayoutModel::class)->findByPk($dc->id))) {
+        if (!$request || !$dc || !$this->scopeMatcher->isBackendRequest($request) || !($layout = $this->contaoFramework->getAdapter(LayoutModel::class)->findByPk($dc->id))) {
+            return;
+        }
+
+        $messageAdapter = $this->contaoFramework->getAdapter(Message::class);
+        if ($layout->addEncore) {
+            if ($messageAdapter->hasMessages('huh.encore.error.noEntryPoints')) {
+                $messageAdapter->addError($messageAdapter->generateUnwrapped('huh.encore.error.noEntryPoints', true));
+            } else {
+                try {
+                    $this->entryCollection->getEntries();
+                } catch (NoEntrypointsException $e) {
+                    $messageAdapter->addError('[Encore Bundle] '.$this->translator->trans('huh.encore.errors.noEntrypoints').' '.$e->getMessage());
+                }
+            }
+        }
+
+        if (!isset($this->bundleConfig['use_contao_template_variables'])
+            || true !== $this->bundleConfig['use_contao_template_variables']) {
             return;
         }
 
         if ($layout->addEncore && $layout->addJQuery && (!isset($this->bundleConfig['unset_jquery']) || true !== $this->bundleConfig['unset_jquery'])) {
-            $this->contaoFramework->getAdapter(Message::class)->addInfo(($GLOBALS['TL_LANG']['tl_layout']['INFO']['jquery_order_conflict'] ?: ''));
+            $messageAdapter->addInfo(($GLOBALS['TL_LANG']['tl_layout']['INFO']['jquery_order_conflict'] ?: ''));
+
         }
     }