+++ src/EventListener/Callback/EncoreEntryOptionListener.php
@@ -8,22 +8,32 @@
 
 namespace HeimrichHannot\EncoreBundle\EventListener\Callback;
 
+use Contao\Message;
 use HeimrichHannot\EncoreBundle\Collection\EntryCollection;
+use HeimrichHannot\EncoreBundle\Exception\NoEntrypointsException;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class EncoreEntryOptionListener
 {
-    private EntryCollection $entryCollection;
+    private EntryCollection     $entryCollection;
+    private TranslatorInterface $translator;
 
-    public function __construct(EntryCollection $entryCollection)
+    public function __construct(EntryCollection $entryCollection, TranslatorInterface $translator)
     {
         $this->entryCollection = $entryCollection;
+        $this->translator = $translator;
     }
 
     public function getEntriesAsOptions(): array
     {
         $choices = [];
 
-        $projectEntries = $this->entryCollection->getEntries();
+        try {
+            $projectEntries = $this->entryCollection->getEntries();
+        } catch (NoEntrypointsException $e) {
+            $projectEntries = [];
+            Message::addError('[Encore Bundle] '.$this->translator->trans('huh.encore.errors.noEntrypoints').' '.$e->getMessage(), 'huh.encore.error.noEntryPoints');
+        }
 
         if (empty($projectEntries)) {
             return $choices;