+++ tests/EventListener/Callback/EncoreEntryOptionListenerTest.php
@@ -11,14 +11,16 @@ namespace HeimrichHannot\EncoreBundle\Test\EventListener\Callback;
 use Contao\TestCase\ContaoTestCase;
 use HeimrichHannot\EncoreBundle\Collection\EntryCollection;
 use HeimrichHannot\EncoreBundle\EventListener\Callback\EncoreEntryOptionListener;
+use Symfony\Contracts\Translation\TranslatorInterface;
 
 class EncoreEntryOptionListenerTest extends ContaoTestCase
 {
     public function createTestInstance(array $parameters = []): EncoreEntryOptionListener
     {
         $entryCollection = $parameters['entryCollection'] ?? $this->createMock(EntryCollection::class);
+        $translator = $parameters['translator'] ?? $this->createMock(TranslatorInterface::class);
 
-        return new EncoreEntryOptionListener($entryCollection);
+        return new EncoreEntryOptionListener($entryCollection, $translator);
     }
 
     public function testGetEntriesAsOptions()